Refactor API round #2
