+++ images/onnx-serve-gpu/Dockerfile
@@ -23,17 +23,18 @@ RUN apt-get update -qq && apt-get install -y -q \
 ENV PYTHONPATH "${PYTHONPATH}:/src:/mnt/project"
 
 COPY pkg/workloads/cortex/lib/requirements.txt /src/cortex/lib/requirements.txt
-COPY pkg/workloads/cortex/onnx_serve/requirements.txt /src/cortex/onnx_serve/requirements.txt
+COPY pkg/workloads/cortex/serve/onnx.requirements.txt /src/cortex/serve/requirements.txt
 
 ARG ONNXRUNTIME_VERSION="1.1.0"
 
-RUN pip install -r /src/cortex/lib/requirements.txt && \
-    pip install -r /src/cortex/onnx_serve/requirements.txt && \
+RUN pip install --upgrade pip && \
+    pip install --no-cache-dir -r /src/cortex/lib/requirements.txt && \
+    pip install --no-cache-dir -r /src/cortex/serve/requirements.txt && \
     pip install onnxruntime-gpu==${ONNXRUNTIME_VERSION} && \
     rm -rf /root/.cache/pip*
 
 COPY pkg/workloads/cortex/consts.py /src/cortex
 COPY pkg/workloads/cortex/lib /src/cortex/lib
-COPY pkg/workloads/cortex/onnx_serve /src/cortex/onnx_serve
+COPY pkg/workloads/cortex/serve /src/cortex/serve
 
-ENTRYPOINT ["/src/cortex/onnx_serve/run.sh"]
+ENTRYPOINT ["/src/cortex/serve/run.sh"]
