+++ pkg/workloads/cortex/serve/serve.py
@@ -21,17 +21,17 @@ from flask import Flask, request, jsonify, g
 from flask_api import status
 from waitress import serve
 
-from cortex.lib import util, API, api_utils, metrics
+from cortex.lib import util, api_utils
+from cortex.lib.type import API
 from cortex.lib.log import cx_logger, debug_obj, refresh_logger
 from cortex.lib.storage import S3
-from cortex.lib.exceptions import UserRuntimeException, UserException, CortexException
-from cortex.tf_api.client import TensorFlowClient
+from cortex.lib.exceptions import UserRuntimeException
 
 app = Flask(__name__)
 app.json_encoder = util.json_tricks_encoder
 
 
-local_cache = {"api": None, "client": None, "class_set": set()}
+local_cache = {"api": None, "predictor_impl": None, "client": None, "class_set": set()}
 
 
 @app.before_request
@@ -57,7 +57,16 @@ def after_request(response):
     if "prediction" in g:
         prediction = g.prediction
 
-    metrics.post_request_metrics(api, response, prediction, g.start_time, local_cache["class_set"])
+    try:
+        predicted_value = None
+        if api.tracker is not None:
+            predicted_value = api.tracker.extract_predicted_value(prediction)
+            api.post_request_metrics(response.status_code, g.start_time, predicted_value)
+            if predicted_value is not None and predicted_value not in local_cache["class_set"]:
+                api.upload_class(predicted_value)
+                local_cache["class_set"].add(predicted_value)
+    except Exception as e:
+        cx_logger().warn("unable to record prediction metric", exc_info=True)
 
     return response
 
@@ -72,12 +81,12 @@ def predict():
         return "malformed json", status.HTTP_400_BAD_REQUEST
 
     api = local_cache["api"]
-    predictor = local_cache["predictor"]
+    predictor_impl = local_cache["predictor_impl"]
 
     try:
         debug_obj("payload", payload, debug)
         try:
-            output = predictor.predict(payload)
+            output = predictor_impl.predict(payload)
         except Exception as e:
             raise UserRuntimeException(api.predictor.path, "predict", str(e)) from e
         debug_obj("prediction", output, debug)
@@ -97,58 +106,11 @@ def prediction_failed(reason):
 
 @app.route("/predict", methods=["GET"])
 def get_summary():
-    response = {
-        "model_signature": local_cache["client"].input_signature,
-        "message": api_utils.API_SUMMARY_MESSAGE,
-    }
+    response = {"message": api_utils.API_SUMMARY_MESSAGE}
+
+    if hasattr(local_cache["client"], "input_signature"):
+        response["model_signature"] = local_cache["client"].input_signature
     return jsonify(response)
-
-
-tf_expected_dir_structure = """tensorflow model directories must have the following structure:
-  1523423423/ (version prefix, usually a timestamp)
-  ├── saved_model.pb
-  └── variables/
-      ├── variables.index
-      ├── variables.data-00000-of-00003
-      ├── variables.data-00001-of-00003
-      └── variables.data-00002-of-...`"""
-
-
-def validate_model_dir(model_dir):
-    version = None
-    for file_name in os.listdir(model_dir):
-        if file_name.isdigit():
-            version = file_name
-            break
-
-    if version is None:
-        cx_logger().error(tf_expected_dir_structure)
-        raise UserException("no top-level version folder found")
-
-    if not os.path.isdir(os.path.join(model_dir, version)):
-        cx_logger().error(tf_expected_dir_structure)
-        raise UserException("no top-level version folder found")
-
-    if not os.path.isfile(os.path.join(model_dir, version, "saved_model.pb")):
-        cx_logger().error(tf_expected_dir_structure)
-        raise UserException('expected a "saved_model.pb" file')
-
-    if not os.path.isdir(os.path.join(model_dir, version, "variables")):
-        cx_logger().error(tf_expected_dir_structure)
-        raise UserException('expected a "variables" directory')
-
-    if not os.path.isfile(os.path.join(model_dir, version, "variables", "variables.index")):
-        cx_logger().error(tf_expected_dir_structure)
-        raise UserException('expected a "variables/variables.index" file')
-
-    for file_name in os.listdir(os.path.join(model_dir, version, "variables")):
-        if file_name.startswith("variables.data-00000-of"):
-            return
-
-    cx_logger().error(tf_expected_dir_structure)
-    raise UserException(
-        'expected at least one variables data file, starting with "variables.data-00000-of-"'
-    )
 
 
 @app.errorhandler(Exception)
@@ -163,43 +125,22 @@ def start(args):
     try:
         raw_api_spec = api_utils.get_spec(args.cache_dir, args.spec)
         api = API(storage=storage, cache_dir=args.cache_dir, **raw_api_spec)
-        local_cache["api"] = api
-
-        if api.predictor.type != "tensorflow":
-            raise CortexException(api.name, "predictor type is not tensorflow")
-
-        local_cache["client"] = TensorFlowClient(
-            "localhost:" + str(args.tf_serve_port), api.predictor.signature_key
-        )
-
+        client = api.predictor.initialize_client(args)
         cx_logger().info("loading the predictor from {}".format(api.predictor.path))
+        predictor_impl = api.predictor.initialize_impl(args.project_dir, client)
 
-        predictor_class = api.predictor.class_impl(args.project_dir)
-
-        try:
-            local_cache["predictor"] = predictor_class(local_cache["client"], api.predictor.config)
-        except Exception as e:
-            raise UserRuntimeException(api.predictor.path, "__init__", str(e)) from e
-        finally:
-            refresh_logger()
-
-    except Exception as e:
+        local_cache["api"] = api
+        local_cache["client"] = client
+        local_cache["predictor_impl"] = predictor_impl
+    except:
         cx_logger().exception("failed to start api")
-        sys.exit(1)
-
-    try:
-        validate_model_dir(args.model_dir)
-    except Exception as e:
-        cx_logger().exception("failed to validate model")
         sys.exit(1)
 
     if api.tracker is not None and api.tracker.model_type == "classification":
         try:
-            local_cache["class_set"] = metrics.get_classes(api)
+            local_cache["class_set"] = api.get_cached_classes()
         except Exception as e:
             cx_logger().warn("an error occurred while attempting to load classes", exc_info=True)
-
-    cx_logger().info("TensorFlow model signature: {}".format(local_cache["client"].input_signature))
 
     waitress_kwargs = api_utils.extract_waitress_params(api.predictor.config)
     waitress_kwargs["listen"] = "*:{}".format(args.port)
@@ -214,12 +155,15 @@ def main():
     na = parser.add_argument_group("required named arguments")
     na.add_argument("--port", type=int, required=True, help="port (on localhost) to use")
     na.add_argument(
-        "--tf-serve-port", type=int, required=True, help="port (on localhost) where tf serving runs"
+        "--tf-serve-port",
+        type=int,
+        required=False,
+        help="port (on localhost) where tf serving runs",
     )
     na.add_argument(
         "--spec", required=True, help="s3 path to api spec (e.g. s3://bucket/path/to/api_spec.json)"
     )
-    na.add_argument("--model-dir", required=True, help="directory to download the model to")
+    na.add_argument("--model-dir", required=False, help="directory to download the model to")
     na.add_argument("--cache-dir", required=True, help="local path for the api cache")
     na.add_argument("--project-dir", required=True, help="local path for the project zip file")
     parser.set_defaults(func=start)
